Enter keybind command
The Enter keybind command is now the appropriate one for the active widget. So for the Set Card Quantity entry it sets the card quantity, and for the Filter options it filters the deck list. 
Also fixed an issue where it was loading the card database a bunch of times unnecessarily.