Bug fix
Decklist search results weren't being displayed. They now show up in their own tab and can be closed with a button. Also added a button to make the default sorts of the Cards in a Deck if that hasn't already been done, like if there was an old .mdk file.